requirements.txt
asgiref==3.8.1
Django==5.0.6
django-environ==0.11.2
djangorestframework==3.15.1
mysqlclient==2.2.4
sqlparse==0.5.0
typing_extensions==4.12.0
django-filter==24.2

